patch() fix to support tuples again
The last update to patch() to take both tuples and lists of tuples wasn't tested, and it turns out that a tuple input needs explicitly casting to a single item list in order to be parsed correctly. This is now handled, as well as raising an exception for true TypeErrors.